requirements.txt
requests
pypdf
python-dotenv
python-docx
pytesseract
Pillow
